# PoC of issue #1830