https://github.com/stylelint/stylelint/issues/1830
/cc @davidtheclark @jeddy3 
BTW, we not ignore semicolon after close brace as we wanted to do originally :smile: 
Maybe be we will add `counter` (maybe best name) to `style-search`, it is reduce code.