@evilebottnawi 2016-09-20
Also improved benchmark.
Before:
```shell
Warnings: 0
Mean: 62.5059776849315 ms
Deviation: 7.14222910936161 ms
```
After:
```shell
Warnings: 0
Mean: 35.90618447959182 ms
Deviation: 8.182562908435182 ms
```