@jeddy3 2016-09-20
Changelog entry here: https://github.com/stylelint/stylelint/wiki/Changelog-updates